Added daemon balances to withdraw request screen

+++ Web/xdn-voting/bot/botUtils.go
@@ -118,7 +118,7 @@ type ActivityBotStruct struct {
 	Id        int    `db:"id"`
 	IdPost    int    `db:"idPost"`
 	IdChannel int64  `db:"idChannel"`
-	IdMessage int64  `db:"idMessage"`
+	IdMessage int    `db:"idMessage"`
 	DateTime  string `db:"dateTime"`
 }
